Extract tentacle core helpers
